requirements.txt
python-telegram-bot
requests
huggingface_hub
python-dotenv